+++ src/lang/lang.mli
@@ -41,6 +41,7 @@ module Ground : sig
 
   type content = Lang_values.Ground.content = {
     descr : unit -> string;
+    to_json : compact:bool -> json5:bool -> unit -> string;
     compare : t -> int;
     typ : Lang_types.ground;
   }
@@ -333,12 +334,14 @@ module type Abstract = sig
   val t : t
   val to_value : content -> value
   val of_value : value -> content
+  val is_value : value -> bool
 end
 
 module type AbstractDef = sig
   type content
 
   val name : string
+  val to_json : compact:bool -> json5:bool -> content -> string
   val descr : content -> string
   val compare : content -> content -> int
 end